+++ CHANGELOG.md
@@ -60,6 +60,15 @@ Master list of UniV3 forks:
 * Add SolidlyV3 UniV3 fork to Sonic
 * Add Wagmi UniV3 fork to Sonic
 
+## 2024-12-18
+
+### Breaking changes
+
+### Non-breaking changes
+
+* Deploy Settler to Ink chain
+  * Add UniswapV3 UniV3 fork to Ink
+
 ## 2024-12-14
 
 ### Breaking changes